feat: implement more static directory node operations fix: show static subdirs

+++ internal/fuse/nodes/static_directory_node.go
@@ -23,13 +23,33 @@ type staticDirectoryNode struct {
 	nodeConfig config.DirectoryConfig
 }
 
+func (n *staticDirectoryNode) isSubdir(name string) bool {
+	for _, subdir := range n.nodeConfig.Subdirectories {
+		if subdir.Name == name {
+			return true
+		}
+	}
+	return false
+}
+
 func (n *staticDirectoryNode) Readdir(ctx context.Context) (fs.DirStream, syscall.Errno) {
-	nodes, err := db.Get().GetNodesForDir(ctx, n.nodeConfig.Tags) // todo: also show static dirs in this list
+	nodes, err := db.Get().GetNodesForDir(ctx, n.nodeConfig.Tags)
 	if err != nil {
-		staticDirLogger.Printf("Readdir: GetFilesForDir failed for %s: %v", n.nodeConfig.Name, err)
+		staticDirLogger.Printf("Readdir: GetNodesForDir failed for %s: %v", n.nodeConfig.Name, err)
 		return nil, syscall.EIO
 	}
-	var result []fuse.DirEntry
+
+	result := make([]fuse.DirEntry, 0, len(nodes)+len(n.nodeConfig.Subdirectories))
+
+	for name, child := range n.Children() {
+		if n.isSubdir(name) {
+			result = append(result, fuse.DirEntry{
+				Name: name,
+				Ino:  child.StableAttr().Ino,
+				Mode: child.Mode(),
+			})
+		}
+	}
 	for _, node := range nodes {
 		result = append(result, fuse.DirEntry{
 			Name: node.OrigName,
@@ -41,9 +61,13 @@ func (n *staticDirectoryNode) Readdir(ctx context.Context) (fs.DirStream, syscal
 }
 
 func (n *staticDirectoryNode) Lookup(ctx context.Context, name string, out *fuse.EntryOut) (*fs.Inode, syscall.Errno) {
+	if child := n.GetChild(name); child != nil {
+		return child, fs.OK
+	}
+
 	nodes, err := db.Get().GetNodesForDir(ctx, n.nodeConfig.Tags)
 	if err != nil {
-		staticDirLogger.Printf("Lookup: GetFilesForDir failed for %s: %v", n.nodeConfig.Name, err)
+		staticDirLogger.Printf("Lookup: GetNodesForDir failed for %s: %v", n.nodeConfig.Name, err)
 		return nil, syscall.EIO
 	}
 
@@ -74,6 +98,9 @@ func (n *staticDirectoryNode) Lookup(ctx context.Context, name string, out *fuse
 func (n *staticDirectoryNode) Mkdir(ctx context.Context, name string, mode uint32, _ *fuse.EntryOut) (*fs.Inode, syscall.Errno) {
 	if !n.nodeConfig.Rules.AllowSubdirCreation {
 		return nil, syscall.EPERM
+	}
+	if child := n.GetChild(name); child != nil {
+		return nil, syscall.EEXIST
 	}
 	staticDirLogger.Printf("Ingesting new directory at %s: %s", n.nodeConfig.Name, name)
 	dirID := ulid.Make().String()
@@ -108,16 +135,19 @@ func (n *staticDirectoryNode) Mkdir(ctx context.Context, name string, mode uint3
 }
 
 func (n *staticDirectoryNode) Rmdir(ctx context.Context, name string) syscall.Errno {
+	if child := n.GetChild(name); child != nil && n.isSubdir(name) {
+		return syscall.EPERM
+	}
 	nodes, err := db.Get().GetNodesForDir(ctx, n.nodeConfig.Tags)
 	if err != nil {
-		staticDirLogger.Printf("Rmdir: GetFilesForDir failed for %s: %v", n.nodeConfig.Name, err)
+		staticDirLogger.Printf("Rmdir: GetNodesForDir failed for %s: %v", n.nodeConfig.Name, err)
 		return syscall.EIO
 	}
 	for _, node := range nodes {
 		if node.OrigName != name {
 			continue
 		}
-		physicalPath := filepath.Join(config.Get().StoragePath, ".data", node.ID, node.OrigName)
+		physicalPath := filepath.Join(config.Get().StoragePath, ".data", node.ID)
 		if err := os.RemoveAll(physicalPath); err != nil {
 			staticDirLogger.Printf("Rmdir: RemoveAll failed for %s: %v", physicalPath, err)
 			return syscall.EIO
@@ -126,15 +156,77 @@ func (n *staticDirectoryNode) Rmdir(ctx context.Context, name string) syscall.Er
 		if err != nil {
 			return fs.ToErrno(err)
 		}
+		return fs.OK
 	}
-	return fs.OK
+	return syscall.ENOENT
 }
 
 func (n *staticDirectoryNode) Rename(ctx context.Context, name string, newParent fs.InodeEmbedder, newName string, flags uint32) syscall.Errno {
-	// todo: implement rename logic for static directories
-	// when in same directory, update name in db and rename physical file
-	// when in different directory, create new node in new directory and delete old node
-	panic("not implemented")
+	if flags != 0 {
+		// RENAME_NOREPLACE / RENAME_EXCHANGE aren't supported by the tag-based store yet.
+		return syscall.ENOSYS
+	}
+	if child := n.GetChild(name); child != nil && n.isSubdir(name) {
+		return syscall.EPERM
+	}
+
+	nodes, err := db.Get().GetNodesForDir(ctx, n.nodeConfig.Tags)
+	if err != nil {
+		staticDirLogger.Printf("Rename: GetNodesForDir failed for %s: %v", n.nodeConfig.Name, err)
+		return syscall.EIO
+	}
+
+	var target gen.Node
+	found := false
+	for _, node := range nodes {
+		if node.OrigName == name {
+			target = node
+			found = true
+			break
+		}
+	}
+	if !found {
+		return syscall.ENOENT
+	}
+
+	newParentNode, ok := newParent.(*staticDirectoryNode)
+	if !ok {
+		staticDirLogger.Printf("Rename: unsupported destination type for %s -> %s", name, newName)
+		return syscall.EXDEV
+	}
+	if newParentNode.GetChild(newName) != nil {
+		return syscall.EEXIST
+	}
+
+	oldPhysicalPath := filepath.Join(config.Get().StoragePath, ".data", target.ID, target.OrigName)
+	newPhysicalPath := filepath.Join(config.Get().StoragePath, ".data", target.ID, newName)
+	if oldPhysicalPath != newPhysicalPath {
+		if err := os.Rename(oldPhysicalPath, newPhysicalPath); err != nil {
+			staticDirLogger.Printf("Rename: physical rename failed %s -> %s: %v", oldPhysicalPath, newPhysicalPath, err)
+			return fs.ToErrno(err)
+		}
+	}
+
+	if err := db.Get().Queries.RenameNode(ctx, gen.RenameNodeParams{
+		ID:       target.ID,
+		OrigName: newName,
+	}); err != nil {
+		staticDirLogger.Printf("Rename: DB rename failed for %s: %v", target.ID, err)
+		if oldPhysicalPath != newPhysicalPath {
+			_ = os.Rename(newPhysicalPath, oldPhysicalPath)
+		}
+		return syscall.EIO
+	}
+
+	if newParentNode.nodeConfig.Name != n.nodeConfig.Name {
+		tags := append(newParentNode.nodeConfig.Tags, logic.GetImplicitTags(newParentNode.nodeConfig.Tags)...)
+		if err := db.Get().UpdateTags(target.ID, tags); err != nil {
+			staticDirLogger.Printf("Rename: UpdateTags failed for %s: %v", target.ID, err)
+			return syscall.EIO
+		}
+	}
+
+	return fs.OK
 }
 
 func (n *staticDirectoryNode) Create(ctx context.Context, name string, flags uint32, mode uint32, _ *fuse.EntryOut) (node *fs.Inode, fh fs.FileHandle, fuseFlags uint32, errno syscall.Errno) {
@@ -157,19 +249,22 @@ func (n *staticDirectoryNode) Create(ctx context.Context, name string, flags uin
 		Mode:     logic.ToStoredMode(mode, false),
 	})
 	if err != nil {
-		err := f.Close()
-		if err != nil {
-			rootLogger.Printf("Error closing physical file: %v", err)
+		if closeErr := f.Close(); closeErr != nil {
+			rootLogger.Printf("Error closing physical file: %v", closeErr)
 			return nil, nil, 0, 0
 		}
-		err = os.Remove(physicalPath)
-		if err != nil {
-			rootLogger.Printf("Error removing physical file: %v", err)
+		if rmErr := os.Remove(physicalPath); rmErr != nil {
+			rootLogger.Printf("Error removing physical file: %v", rmErr)
 			return nil, nil, 0, 0
 		}
 		rootLogger.Printf("Error inserting file into DB: %v", err)
 		return nil, nil, 0, syscall.EIO
 	}
+
+	if err := db.Get().UpdateTags(fileID, n.nodeConfig.Tags); err != nil {
+		staticDirLogger.Printf("Create: UpdateTags failed for %s: %v", fileID, err)
+	}
+
 	childNode := &passthroughNode{Path: physicalPath}
 	childInode := n.NewPersistentInode(ctx, childNode, fs.StableAttr{Mode: mode})
 	return childInode, f, 0, fs.OK
@@ -178,7 +273,7 @@ func (n *staticDirectoryNode) Create(ctx context.Context, name string, flags uin
 func (n *staticDirectoryNode) Open(ctx context.Context, flags uint32) (fh fs.FileHandle, fuseFlags uint32, errno syscall.Errno) {
 	nodes, err := db.Get().GetNodesForDir(ctx, n.nodeConfig.Tags)
 	if err != nil {
-		staticDirLogger.Printf("Lookup: GetFilesForDir failed for %s: %v", n.nodeConfig.Name, err)
+		staticDirLogger.Printf("Open: GetNodesForDir failed for %s: %v", n.nodeConfig.Name, err)
 		return nil, 0, syscall.EIO
 	}
 	for _, node := range nodes {
@@ -194,25 +289,84 @@ func (n *staticDirectoryNode) Open(ctx context.Context, flags uint32) (fh fs.Fil
 	return nil, 0, syscall.ENOENT
 }
 
+func (n *staticDirectoryNode) Opendir(context.Context) syscall.Errno {
+	return fs.OK
+}
+
+func (n *staticDirectoryNode) Access(context.Context, uint32) syscall.Errno {
+	return fs.OK
+}
+
 func (n *staticDirectoryNode) Unlink(ctx context.Context, name string) syscall.Errno {
+	if child := n.GetChild(name); child != nil && n.isSubdir(name) {
+		return syscall.EPERM
+	}
 	nodes, err := db.Get().GetNodesForDir(ctx, n.nodeConfig.Tags)
 	if err != nil {
-		staticDirLogger.Printf("Lookup: GetFilesForDir failed for %s: %v", n.nodeConfig.Name, err)
+		staticDirLogger.Printf("Unlink: GetNodesForDir failed for %s: %v", n.nodeConfig.Name, err)
 		return syscall.EIO
 	}
 	for _, node := range nodes {
 		if node.OrigName == name {
-			physicalPath := filepath.Join(config.Get().StoragePath, ".data", node.ID, node.OrigName)
-			if err := os.Remove(physicalPath); err != nil {
+			physicalPath := filepath.Join(config.Get().StoragePath, ".data", node.ID)
+			if err := os.RemoveAll(physicalPath); err != nil {
 				return fs.ToErrno(err)
 			}
-			err = db.Get().Queries.DeleteNode(ctx, node.ID)
-			if err != nil {
+			if err := db.Get().Queries.DeleteNode(ctx, node.ID); err != nil {
 				return fs.ToErrno(err)
 			}
+			return fs.OK
 		}
 	}
 	return syscall.ENOENT
+}
+
+func (n *staticDirectoryNode) Symlink(ctx context.Context, target, name string, _ *fuse.EntryOut) (*fs.Inode, syscall.Errno) {
+	if !n.nodeConfig.Rules.AllowFileCreation {
+		return nil, syscall.EPERM
+	}
+	if n.GetChild(name) != nil {
+		return nil, syscall.EEXIST
+	}
+	linkID := ulid.Make().String()
+	physicalPath := filepath.Join(config.Get().StoragePath, ".data", linkID, name)
+	if err := os.MkdirAll(filepath.Dir(physicalPath), 0755); err != nil {
+		return nil, fs.ToErrno(err)
+	}
+	if err := os.Symlink(target, physicalPath); err != nil {
+		return nil, fs.ToErrno(err)
+	}
+
+	var st syscall.Stat_t
+	if err := syscall.Lstat(physicalPath, &st); err != nil {
+		return nil, fs.ToErrno(err)
+	}
+
+	err := db.Get().Queries.InsertNode(ctx, gen.InsertNodeParams{
+		ID:       linkID,
+		OrigName: name,
+		Mode:     logic.ToStoredMode(uint32(syscall.S_IFLNK|0777), false),
+	})
+	if err != nil {
+		_ = os.Remove(physicalPath)
+		staticDirLogger.Printf("Symlink: InsertNode failed for %s: %v", name, err)
+		return nil, syscall.EIO
+	}
+	if err := db.Get().UpdateTags(linkID, n.nodeConfig.Tags); err != nil {
+		staticDirLogger.Printf("Symlink: UpdateTags failed for %s: %v", linkID, err)
+	}
+
+	childNode := &passthroughNode{Path: physicalPath}
+	childInode := n.NewPersistentInode(ctx, childNode, fs.StableAttr{Mode: syscall.S_IFLNK})
+	return childInode, fs.OK
+}
+
+func (n *staticDirectoryNode) Link(context.Context, fs.InodeEmbedder, string, *fuse.EntryOut) (node *fs.Inode, errno syscall.Errno) {
+	return nil, syscall.EPERM
+}
+
+func (n *staticDirectoryNode) Mknod(_ context.Context, _ string, _, _ uint32, _ *fuse.EntryOut) (*fs.Inode, syscall.Errno) {
+	return nil, syscall.EPERM
 }
 
 func (n *staticDirectoryNode) Setattr(context.Context, fs.FileHandle, *fuse.SetAttrIn, *fuse.AttrOut) syscall.Errno {
@@ -231,6 +385,11 @@ var _ fs.NodeRmdirer = (*staticDirectoryNode)(nil)
 var _ fs.NodeRenamer = (*staticDirectoryNode)(nil)
 var _ fs.NodeCreater = (*staticDirectoryNode)(nil)
 var _ fs.NodeOpener = (*staticDirectoryNode)(nil)
+var _ fs.NodeOpendirer = (*staticDirectoryNode)(nil)
+var _ fs.NodeAccesser = (*staticDirectoryNode)(nil)
 var _ fs.NodeUnlinker = (*staticDirectoryNode)(nil)
+var _ fs.NodeSymlinker = (*staticDirectoryNode)(nil)
+var _ fs.NodeLinker = (*staticDirectoryNode)(nil)
+var _ fs.NodeMknoder = (*staticDirectoryNode)(nil)
 var _ fs.NodeSetattrer = (*staticDirectoryNode)(nil)
 var _ fs.NodeGetattrer = (*staticDirectoryNode)(nil)
